woodz_bot/requirements.txt
flask==3.0.3
discord.py==2.4.0